fix: #1035 컬렉션/아카이브 어드민을 CMS 편입 기준으로 정리
컬렉션과 아카이브 콘텐츠가 CMS page_blocks로 편입되어 별도 어드민 CRUD가 중복 관리면이 되었다. 관리자 메뉴와 admin API/page를 제거하고, 상품 필터가 사용하는 public collections 조회만 유지한다.

영어 archive CMS 블록에서 작가 CTA 라벨이 한국어로 남는 문제는 page_blocks content의 hrefLabel_en 데이터 백필로 보정한다. e2e는 제거된 admin CRUD 대신 public collections 조회가 유지되는지를 검증하도록 바꿨다.

Constraint: 상품 필터는 public /collections 엔드포인트를 계속 사용함
Constraint: archive/collection 화면 콘텐츠는 CMS page_blocks가 소스임
Rejected: admin collections/archives CRUD를 숨김 처리만 하기 | 중복 API와 테스트가 계속 남아 CMS 전환 의도를 흐림
Confidence: high
Scope-risk: moderate
Directive: 컬렉션/아카이브 전용 어드민을 되살리기 전에 CMS page_blocks 소스와 운영 플로우가 분리되어야 하는지 재검증할 것
Tested: bash scripts/codex-project-guard.sh
Tested: npm run build && npm run test:run
Tested: cd backend && npm run build && npm run test
Tested: bash scripts/test.sh e2e
Tested: runtime smoke /api/health, /ko, /en/collection, /en/archive, /api/collections/clay?locale=en
Not-tested: direct cd backend && npm run test:e2e without wrapper because DATABASE_URL/TEST_DATABASE_URL is intentionally supplied by scripts/test.sh
