+++ backend/src/modules/collections/collections.module.ts
@@ -2,12 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Collection } from './entities/collection.entity';
 import { CollectionsController } from './collections.controller';
-import { AdminCollectionsController } from './admin-collections.controller';
 import { CollectionsService } from './collections.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Collection])],
-  controllers: [CollectionsController, AdminCollectionsController],
+  controllers: [CollectionsController],
   providers: [CollectionsService],
   exports: [CollectionsService],
 })
