+++ src/i18n/messages/en.json
@@ -973,9 +973,7 @@
       "settingsGroup": "Site Settings",
       "inquiries": "Inquiries",
       "announcementBars": "Announcement Bars",
-      "collections": "Collections",
       "journal": "Journal",
-      "archives": "Archives",
       "business": "Business Info",
       "close": "Close sidebar",
       "cmsGroup": "CMS",
