+++ src/i18n/messages/ko.json
@@ -973,9 +973,7 @@
       "settingsGroup": "사이트 설정",
       "inquiries": "문의관리",
       "announcementBars": "안내바관리",
-      "collections": "컬렉션관리",
       "journal": "저널관리",
-      "archives": "아카이브관리",
       "business": "사업자 정보",
       "close": "사이드바 닫기",
       "cmsGroup": "CMS",
