+++ src/utils/toastMessages.ts
@@ -74,20 +74,12 @@ const TOAST_MESSAGES = {
     visible: '표시',
     hidden: '숨김',
     settingsSaveError: '설정 저장에 실패했습니다.',
-    clayTypeUpdated: '니료 타입이 수정되었습니다.',
-    clayTypeCreated: '니료 타입이 추가되었습니다.',
-    processUpdated: '공정이 수정되었습니다.',
-    processCreated: '공정이 추가되었습니다.',
-    artistUpdated: '아티스트가 수정되었습니다.',
-    artistCreated: '아티스트가 추가되었습니다.',
     faqQuestionAnswerRequired: '질문과 답변을 입력해주세요.',
     faqUpdated: 'FAQ가 수정되었습니다.',
     faqCreated: 'FAQ가 생성되었습니다.',
     faqSaveError: 'FAQ 저장에 실패했습니다.',
     faqDeleted: 'FAQ가 삭제되었습니다.',
     faqDeleteError: 'FAQ 삭제에 실패했습니다.',
-    collectionUpdated: '컬렉션이 수정되었습니다.',
-    collectionCreated: '컬렉션이 추가되었습니다.',
     categoryUpdated: '카테고리가 수정되었습니다.',
     categoryCreated: '카테고리가 추가되었습니다.',
     sameLevelReorderOnly: '같은 레벨에서만 순서를 변경할 수 있습니다.',
@@ -176,20 +168,12 @@ const TOAST_MESSAGES = {
     visible: 'visible',
     hidden: 'hidden',
     settingsSaveError: 'Failed to save settings.',
-    clayTypeUpdated: 'Clay type updated.',
-    clayTypeCreated: 'Clay type created.',
-    processUpdated: 'Process updated.',
-    processCreated: 'Process created.',
-    artistUpdated: 'Artist updated.',
-    artistCreated: 'Artist created.',
     faqQuestionAnswerRequired: 'Please enter both question and answer.',
     faqUpdated: 'FAQ updated.',
     faqCreated: 'FAQ created.',
     faqSaveError: 'Failed to save FAQ.',
     faqDeleted: 'FAQ deleted.',
     faqDeleteError: 'Failed to delete FAQ.',
-    collectionUpdated: 'Collection updated.',
-    collectionCreated: 'Collection created.',
     categoryUpdated: 'Category updated.',
     categoryCreated: 'Category created.',
     sameLevelReorderOnly: 'You can only reorder items at the same level.',
